+++ bgjobs/tests.py
@@ -1,3 +1,5 @@
-from django.test import TestCase
+"""Tests for the bgjobs app."""
+
+# from django.test import TestCase
 
 # Create your tests here.