+++ bgjobs/views.py
@@ -5,6 +5,7 @@ from django.urls import reverse
 from django.views.generic import TemplateView, ListView
 
 from projectroles.views import LoggedInPermissionMixin, ProjectContextMixin, ProjectPermissionMixin
+from projectroles.plugins import get_backend_api
 
 from .models import BackgroundJob
 
@@ -23,6 +24,7 @@ class ProjectBackgroundJobView(
     model = BackgroundJob
 
     def get_queryset(self):
+        # TODO: filter to user's job if can only see their own
         return super().get_queryset().filter(project__sodar_uuid=self.kwargs["project"])
 
 
@@ -35,8 +37,11 @@ class BackgroundJobClearViewBase(
 ):
     """Base class for view clearing jobs."""
 
+    #: The template is the same for both sub classes.
     template_name = "bgjobs/backgroundjob_confirm_clear.html"
+    #: Set in sub class.
     which_jobs = None
+    #: Set in sub class.
     permission_required = None
 
     def get_context_data(self, *args, **kwargs):
@@ -44,7 +49,8 @@ class BackgroundJobClearViewBase(
         result["which_jobs"] = self.which_jobs
         return result
 
-    def post(self, request, **kwargs):
+    def post(self, _request, **_kwargs):
+        """Handle form POST."""
         context = self.get_context_data()
         project = context["project"]
         filter_kwargs = {"project": project}
@@ -56,12 +62,24 @@ class BackgroundJobClearViewBase(
             bg_job_count = bg_jobs.count()
             bg_jobs.delete()
 
+            timeline = get_backend_api("timeline_backend")
+            if timeline:
+                timeline.add_event(
+                    project=self._get_project(self.request, self.kwargs),
+                    app_name="bgjobs",
+                    user=self.request.user,
+                    event_name="clear_bg_jobs",
+                    description="Clearing {} background jobs".format(
+                        "user-owned" if self.which_jobs != "all" else "all"
+                    ),
+                    status_type="OK",
+                )
             messages.success(self.request, "Removed {} background jobs".format(bg_job_count))
         except Exception as ex:
             messages.error(self.request, "Unable to remove background jobs: {}".format(ex))
 
         return HttpResponseRedirect(
-            reverse("bgjobs:job-list", kwargs={"project": project.sodar_uuid})
+            reverse("bgjobs:jobs-list", kwargs={"project": project.sodar_uuid})
         )